fix: 게이트웨이 코드 수정
