+++ gateway/build.gradle
@@ -9,11 +9,12 @@ dependencies {
 	runtimeOnly group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.11.5'
 
 	implementation 'org.springframework.boot:spring-boot-starter-actuator'
-	implementation 'org.springframework.cloud:spring-cloud-starter-gateway-mvc'
+	implementation 'org.springframework.cloud:spring-cloud-starter-gateway'   // 비동기적
+//	implementation 'org.springframework.cloud:spring-cloud-starter-gateway-mvc'  //  동기적  - 특별한 일 없으면 쓸 일 없음
 	implementation 'org.springframework.cloud:spring-cloud-starter-loadbalancer'
 
 	implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
-	implementation 'org.springframework.boot:spring-boot-starter-web'
+//	implementation 'org.springframework.boot:spring-boot-starter-web'
 }
 
 dependencyManagement {
